requirements.txt
# Multi-Agent Negotiation Simulator — Dependencies

# Web framework & server
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# AI layer
openai>=1.30.0

# Data validation
pydantic>=2.7.0

# Environment / config
python-dotenv>=1.0.0

# Test utilities
httpx>=0.27.0
websockets>=12.0
